move determinstic stuff to separate folder, remove WO dumps
