+++ QSB/Syncs/Occasional/OccasionalManager.cs
@@ -1,6 +1,7 @@
 ﻿using Cysharp.Threading.Tasks;
 using Mirror;
 using QSB.Utility;
+using QSB.Utility.Deterministic;
 using QSB.WorldSync;
 using System.Collections.Generic;
 using System.Linq;
